se modifica listado, se agrega busqueda, se agregan alarmas y validaciones en agregar empleado

+++ src/app/modules/employees/employee-list/employee-list.component.ts
@@ -11,6 +11,10 @@ import { EmployeeService } from '../service/employee.service';
 export class EmployeeListComponent implements OnInit, AfterViewInit {
 
   employees: IEmployee[] = [];
+  filterName: string;
+  filterRfc: string;
+  filterStatus: number;
+  showFilter: boolean;
 
   constructor(
     private employeeService: EmployeeService
@@ -18,15 +22,25 @@ export class EmployeeListComponent implements OnInit, AfterViewInit {
 
 
   ngAfterViewInit(): void {
-    this.getEmployees();
+    this.getEmployees(0);
   }
 
   ngOnInit(): void {
-    this.getEmployees();
+    this.showFilter = false;
   }
 
-  getEmployees(): void {
-    this.employeeService.getEmployees().subscribe((response: IEmployee[]) => {
+  getEmployees(consult: number): void {
+
+    const params = {
+      typeConsult: consult,
+      status: this.filterStatus,
+      name: this.filterName,
+      rfc: this.filterRfc
+    }
+
+    console.log(params);
+
+    this.employeeService.getEmployees(params).subscribe((response: IEmployee[]) => {
       this.employees = response;
       this.employees.map(employee => {
         employee.fullName = `${employee.firstName} ${employee.lastName} ${employee.middleName}`
@@ -37,14 +51,24 @@ export class EmployeeListComponent implements OnInit, AfterViewInit {
   }
 
   onRefreshList(): void {
-    this.getEmployees();
+    this.showFilter = false;
+    this.filterName = '';
+    this.filterRfc = '';
+    this.filterStatus = 0;
+    this.getEmployees(0);
   }
 
   onRemoveEmployee(id: string): void {
-    this.employeeService.deleteEmployee(id).subscribe((respose:IEmployee)=>{
-      this.getEmployees();
-    }, (error: HttpErrorResponse) => {
-      console.error(error);
-    });
+    if (confirm(`¿Esta seguro de eliminar el empleado?`)) {
+      this.employeeService.deleteEmployee(id).subscribe((respose: IEmployee) => {
+        this.getEmployees(0);
+      }, (error: HttpErrorResponse) => {
+        alert(error.message);
+      });
+    }
+  }
+
+  onFilterList(): void {
+    this.getEmployees(1)
   }
 }
